+++ internal/backends/qt/qt_window.rs
@@ -1444,14 +1444,12 @@ impl QtWindow {
 
         let text = qt_key_to_string(key as key_generated::Qt_Key, text);
 
-        for ch in text.chars() {
-            let event = if released {
-                WindowEvent::KeyReleased { text: ch }
-            } else {
-                WindowEvent::KeyPressed { text: ch }
-            };
-            self.window.dispatch_event(event)
-        }
+        let event = if released {
+            WindowEvent::KeyReleased { text }
+        } else {
+            WindowEvent::KeyPressed { text }
+        };
+        self.window.dispatch_event(event);
 
         timer_event();
     }